API_trading/requirements.txt
Django
djangorestframework==3.15.2
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0
psycopg2
